screenshots: remove redundancy code, switch processing to filters
<==== Mercurial data ====>
Branch: rotate_image_filter

+++ common/src/transcoding/filters/filter_helper.cpp
@@ -59,7 +59,7 @@ void QnImageFilterHelper::setContrastParams(const ImageCorrectionParams& params)
     m_contrastParams = params;
 }
 
-void QnImageFilterHelper::setTimeCorner(Qn::Corner corner, int onscreenDateOffset)
+void QnImageFilterHelper::setTimeCorner(Qn::Corner corner, qint64 onscreenDateOffset)
 {
     m_timestampCorner = corner;
     m_onscreenDateOffset = onscreenDateOffset;
